chore: rename vector-score folder to vector-scores
