+++ webapp/app/(routes)/page.tsx
@@ -1,6 +1,6 @@
 import Resume from "@/components/resume/resume";
 import JobDescriptions from "@/components/job-descriptions/job-descriptions";
-import VectorScore from "@/components/vector-score/vector-scores";
+import VectorScore from "@/components/vector-scores/vector-scores";
 import CommonWords from "@/components/common-words/common-words";
 import Suggestions from "@/components/suggestions/suggestions";
